Refactored updateCommitsTable into a class named CommitTableUseCases
