+++ server/src/modules/Commits/Application/getCommits.ts
@@ -1,25 +1,27 @@
 import { GithubAdapter } from "../../Github/Repositories/github.API";
 import { CommitRepository } from "../Repositories/commitRepository";
-import { getCommitsAPI, saveCommitsDB, checkNewCommits } from "./updateCommitsTable";
+import { CommitTableUseCases } from "./CommitTableUseCases";
 
 export class CommitUseCases {
+  private commitTableUseCases: CommitTableUseCases;
   constructor(
     private repositoryAdapter: CommitRepository,
-    private githubAdapter: GithubAdapter
+    private githubAdapter: GithubAdapter,
   ) {
     this.repositoryAdapter = repositoryAdapter;
     this.githubAdapter = githubAdapter;
+    this.commitTableUseCases = new CommitTableUseCases(this.repositoryAdapter, this.githubAdapter);
   }
 
   async getCommits(owner: string, repoName: string) {
     try {
       if (!await this.repositoryAdapter.repositoryExist(owner, repoName)) {
-        const commits = await getCommitsAPI(owner, repoName, this.githubAdapter);
-        saveCommitsDB(owner, repoName, commits, this.repositoryAdapter);
+        const commits = await this.commitTableUseCases.getCommitsAPI(owner, repoName);
+        this.commitTableUseCases.saveCommitsDB(owner, repoName, commits);
       } else {
-        const commits = await getCommitsAPI(owner, repoName, this.githubAdapter); //getCommitsAPI should be changed to getLastCommits once it is implemented
-        const newCommits = await checkNewCommits(owner, repoName, commits);
-        saveCommitsDB(owner, repoName, newCommits, this.repositoryAdapter);
+        const commits = await this.commitTableUseCases.getCommitsAPI(owner, repoName); //getCommitsAPI should be changed to getLastCommits once it is implemented
+        const newCommits = await this.commitTableUseCases.checkNewCommits(owner, repoName, commits);
+        this.commitTableUseCases.saveCommitsDB(owner, repoName, newCommits);
       }
     } catch (error) {
       console.error("Error updating commits table:", error);
